AST -> Advanced Software Technology, fix GNU troff incompatibility
* README.md: AT&T Software Technology -> Advanced Software Technology
As a reference for the origin of the name, see the abstract of "The
AT&T AST OpenSource Software Collection" (2000) by Fowler, Korn, and
Vo.
* src/cmd/nmake/HISTORY: No reasoning was provided for this 2013 alpha
release change to nmake.mm; it is now explicitly documented as unmerged.
* src/cmd/nmake/nmake.mm: Add missing title and author macros that
were causing errors from GNU troff's version of the MM (AT&T DWB)
macros.  This updated version continues to work with troff2html.