Add .clang-format, format variable.c
The new Clang format file is intended to emulate with the greatest
possible fidelity the style codified within the BARR-C:2018 Embedded C
Coding Standard. This standard is a supplement to the MISRA C
standard, and, like the latter, is intended to reduce the number of
defects when programming in the C language.

Many of the conventions (e.g., opening and closing brackets on their
own lines) of the BARR-C standard were already followed within the AST
source code, but the line lengths, especially those of the conditional
statements, are frequently excessively long (>200) and therefore
difficult to read and maintain.

Although much additional work needs to be done to make the AST more
accessible for newcomers, reformatting is relatively inexpensive
change.