NMAKE: --novirtual when VPATH contains fewer than two directories
NMAKE features extensive support for viewpaths and the creation of
corresponding virtual directory hierarchies, which is the basis for
bin/package's support for out-of-source builds.

When only one viewpath is supplied, this is likely the main source
repository and the creation of directories within would be
unintentional. Therefore, --novirtual is now assumed when the VPATH
contains only a single path.

This patch was back-ported from the 2013-10-10 alpha release of AST.

* src/cmd/nmake/Makerules.mk: Set --novirtual when a second VIEW
argument is non-existent.
* src/cmd/nmake/RELEASE: Include historical revision message.
* src/cmd/nmake/tests/assert.tst
* src/cmd/nmake/tests/option.tst
* src/cmd/nmake/tests/state.tst
* src/cmd/nmake/tests/view.tst: Adjust tests for --novirtual by default.