NMAKE: pkg-ada.mk improvements and external scanner fix
One of the aims for pkg-ada.mk is to reduce to a minimum the reliance
on vendor-supplied tools. The two components of a minimal build of
GNAT, for instance, are the `gnat1` front-end and `gnatbind`. The
remaining GNAT tools must be built subsequently. By dropping the
dependency on `gnatls`, pkg-ada.mk now only requires these two
components.

The NMAKE authors were very concerned with the stability and
reproducibility of builds, which is well-attuned with the demands of
Ada. The -I- option, which disallows inclusion of headers from the
current directory, has been borrowed from the C portion of NMAKE and
is now applied by default with pkg-ada.mk as well. In addition, the
Ada Run-Time System (RTS) path is now explicitly specified, along with
-nostdinc to prevent GNAT from considering the built-in default.

With the modifications to `gnatscan`, this means that NMAKE is now
capable of tracking the modification times and file sizes of the
entire closure of sources (including RTS) within a given Ada
partition.

* src/cmd/nmake/gnatscan:
- sed expressions: use ':' instead of ',' as the delimiter.
- Delete the contents of GNATSCAN_TMPDIR and `rmdir` the directory.
Recursive `rm -f` is inherently dangerous.
- Drop `gnatls` and instead parse the ALI files directly for
  dependencies.
* src/cmd/nmake/pkg-ada.mk:
- Drop `gnatls` and attempt to obtain the adalib and adainclude
  directories directly from `gcc`.